refactor: added @PathVarible in the DELETE request for the PomodoroSettings controller.

+++ backend/src/main/java/com/williamntlam/taskmanagementapp/service/PomodoroSettingsService.java
@@ -1,5 +1,6 @@
 package com.williamntlam.taskmanagementapp.service;
 
+import org.springframework.transaction.annotation.Transactional;
 import com.williamntlam.taskmanagementapp.model.PomodoroSettings;
 import com.williamntlam.taskmanagementapp.repository.PomodoroSettingsRepository;
 import org.springframework.stereotype.Service;
@@ -26,10 +27,13 @@ public class PomodoroSettingsService {
     return pomodoroSettingsRepository.save(settings);
   }
 
+  @Transactional
   public void deleteSettings(Long id) {
-
+    if (!pomodoroSettingsRepository.existsById(id)) {
+        throw new IllegalArgumentException("Pomodoro settings with ID " + id + " do not exist.");
+    }
     pomodoroSettingsRepository.deleteById(id);
-  }
+}
 
   public PomodoroSettings createDefaultSettings(Long userId) {
